Add files via upload
In this project, I analyzed international debt data collected by The World Bank. The dataset contains information about the amount of debt (in USD) owed by developing countries across several categories. Some of the questions I answered are:
*What is the total amount of debt that is owed by the countries listed in the dataset?
*Which country owns the maximum amount of debt and what does that amount look like?
*What is the average amount of debt owed by countries across different debt indicators?
The data used in this project is provided by The World Bank. It contains both national and regional debt statistics for several countries across the globe as recorded from 1970 to 2015.